docs/requirements.txt
graphviz
jupyter-sphinx
sphinx
sphinx-audeering-theme >=1.2.1
sphinx-autodoc-typehints
sphinx-copybutton
sphinxcontrib-katex
